+++ distant-net/src/common.rs
@@ -20,5 +20,4 @@ pub use listener::*;
 pub use map::*;
 pub use packet::*;
 pub use port::*;
-pub use serde_json::Value;
 pub use transport::*;